Move NTP function to ntp package and add ToTime32
Moves all NTP related functions to the NTP package and adds a function
to recover a time.Time from a uint32 encoded NTP timestamp.
