+++ internal/ntp/ntp_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNTPTimeConverstion(t *testing.T) {
+func TestNTPToTimeConverstion(t *testing.T) {
 	for i, cc := range []struct {
 		ts time.Time
 	}{
@@ -24,6 +24,7 @@ func TestNTPTimeConverstion(t *testing.T) {
 	} {
 		t.Run(fmt.Sprintf("TimeToNTP/%v", i), func(t *testing.T) {
 			assert.InDelta(t, cc.ts.UnixNano(), ToTime(ToNTP(cc.ts)).UnixNano(), float64(time.Millisecond.Nanoseconds()))
+			assert.InDelta(t, cc.ts.UnixNano(), ToTime32(ToNTP32(cc.ts), cc.ts).UnixNano(), float64(time.Millisecond.Nanoseconds()))
 		})
 	}
 }
@@ -47,6 +48,41 @@ func TestTimeToNTPConverstion(t *testing.T) {
 	} {
 		t.Run(fmt.Sprintf("TimeToNTP/%v", i), func(t *testing.T) {
 			assert.Equal(t, cc.ts, ToNTP(ToTime(cc.ts)))
+		})
+	}
+}
+
+func TestNTPTime32(t *testing.T) {
+	zero := time.Date(1900, time.January, 1, 0, 0, 0, 0, time.UTC)
+	notSoLongAgo := time.Date(2022, time.May, 5, 14, 48, 20, 0, time.UTC)
+	for i, cc := range []struct {
+		input    time.Time
+		expected uint32
+	}{
+		{
+			input:    zero,
+			expected: 0,
+		},
+		{
+			input:    zero.Add(time.Second),
+			expected: 1 << 16,
+		},
+		{
+			input:    notSoLongAgo,
+			expected: uint32(uint(notSoLongAgo.Sub(zero).Seconds())&0xffff) << 16,
+		},
+		{
+			input:    zero.Add(400 * time.Millisecond),
+			expected: 26214,
+		},
+		{
+			input:    zero.Add(1400 * time.Millisecond),
+			expected: 1<<16 + 26214,
+		},
+	} {
+		t.Run(fmt.Sprintf("%v", i), func(t *testing.T) {
+			res := ToNTP32(cc.input)
+			assert.Equalf(t, cc.expected, res, "%b != %b", cc.expected, res)
 		})
 	}
 }
