+++ pkg/rfc8888/recorder_test.go
@@ -4,7 +4,6 @@
 package rfc8888
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
@@ -160,39 +159,4 @@ func TestRecorder(t *testing.T) {
 			assert.Less(t, 3, len(reports.ReportBlocks[i].MetricBlocks))
 		}
 	})
-}
-
-func TestNTPTime32(t *testing.T) {
-	zero := time.Date(1900, time.January, 1, 0, 0, 0, 0, time.UTC)
-	notSoLongAgo := time.Date(2022, time.May, 5, 14, 48, 20, 0, time.UTC)
-	for i, cc := range []struct {
-		input    time.Time
-		expected uint32
-	}{
-		{
-			input:    zero,
-			expected: 0,
-		},
-		{
-			input:    zero.Add(time.Second),
-			expected: 1 << 16,
-		},
-		{
-			input:    notSoLongAgo,
-			expected: uint32(uint(notSoLongAgo.Sub(zero).Seconds())&0xffff) << 16,
-		},
-		{
-			input:    zero.Add(400 * time.Millisecond),
-			expected: 26214,
-		},
-		{
-			input:    zero.Add(1400 * time.Millisecond),
-			expected: 1<<16 + 26214,
-		},
-	} {
-		t.Run(fmt.Sprintf("%v", i), func(t *testing.T) {
-			res := ntpTime32(cc.input)
-			assert.Equalf(t, cc.expected, res, "%b != %b", cc.expected, res)
-		})
-	}
 }
